% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/linefuncs.R
\name{addprior}
\alias{addprior}
\title{Add a prior to the RLL}
\usage{
addprior(lines, alpha_e = 0, beta_e = 0, alpha_s = 0, beta_s = 0)
}
\arguments{
\item{alpha_e}{a positive numeric. The shape parameter for the IG prior on
\eqn{\sigma^2_e}.}

\item{beta_e}{a positive numeric. The scale parameter for the IG prior on
\eqn{\sigma^2_e}.}

\item{alpha_s}{a positive numeric. The shape parameter for the IG prior on
\eqn{\sigma^2_s}.}

\item{beta_s}{a positive numeric. The scale parameter for the IG prior on
  \eqn{\sigma^2_s}.}
}
\value{
Returns a new \code{lines} dataframe with the constants associated
with the prior term appended to the bottom.
}
\description{
Modifies the \code{lines} dataframe that is output from the
\code{\link{findlines}} function to add a term corresponding to the
prior for use in a Bayesian analysis.
}
\details{
Working with the posterior density of \eqn{\sigma^2_e, \sigma^2_e} instead
of the RLL requires this incorporation of a prior distribution on those same
parameters. We work here with the inverse Gamma, which is the conjugate prior,
\enumerate{
 \item \eqn{\sigma^2_e ~ IG(\alpha_e, \beta_e)}
 \item \eqn{\sigma^2_s ~ IG(\alpha_s, \beta_s)}
 }
Given a dataframe output from \code{findlines},
}

