% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/linefuncs.R
\name{findlines}
\alias{findlines}
\title{Find the constants that define each line.}
\usage{
findlines(x, z, y, SigE, SigS)
}
\arguments{
\item{x}{a matrix with n rows corresponding to the fixed effects.}

\item{z}{a matrix with n rows corresponding to the random effects.}

\item{y}{a numeric vector of length n.}

\item{SigE}{an n x n covariance matrix for the random error.}

\item{SigS}{an n x n covariance matrix for the random effects.}
}
\value{
A dataframe of containing the constants that define the shape of each
  term in the sum.
}
\description{
Given X, Y, Z, find the constants {a_j, b_j, c_j, d_j} that define the shape
of each term in the sum.
}

