% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boxfuncs.R
\name{getbounds}
\alias{getbounds}
\title{Get bounds for a given box}
\usage{
getbounds(lims.sigsqe, lims.sigsqs, lines, status)
}
\arguments{
\item{lims.sigsqe}{numeric vector of length 2 of the form \code{c(lower, upper)}
containing the x-limits of the box.}

\item{lims.sigsqs}{numeric vector of length 2 of the form \code{c(lower, upper)}
containing the y-limits of the box.}

\item{lines}{dataframe containing the constants that define the shape of each
term in the sum. Output from \code{\link{findlines}}.}

\item{status}{factor of length of \code{lines} denoting the location of the box
  relative to each line.}
}
\value{
A matrix with a row corresponding each term in the sum. The first column
contains the lower bound of that term within the box and the second column
contains the upper bound. Note that the \code{colSums} of this matrix yields the
\code{c(lower, upper)} bounds on the full objective function.
}
\description{
Computes the lower and upper bounds on every term in the sum of
the objective function within a box.
}

