% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boxfuncs.R
\name{getstatus}
\alias{getstatus}
\title{Get the location of a box relative to lines.}
\usage{
getstatus(lims.sigsqe, lims.sigsqs, lines)
}
\arguments{
\item{lims.sigsqe}{numeric vector of length 2 of the form \code{c(lower, upper)}
containing the x-limits of the box.}

\item{lims.sigsqs}{numeric vector of length 2 of the form \code{c(lower, upper)}
containing the y-limits of the box.}

\item{lines}{dataframe containing the constants that define the shape of each
  term in the sum. Output from \code{\link{findlines}}.}
}
\value{
A factor of length of \code{lines} denoting the location of the box
  relative to each line.
}
\description{
Computes the location of a box with given limits relative to each of
the lines.
}

