% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{gmst}
\alias{gmst}
\title{Global mean surface temperature}
\format{A data frame with 125 rows and 2 variables:
\describe{
  \item{Year}{integer year}
  \item{temp.dev}{mean temperature deviation for that year (in units 0.01C)}
}}
\source{
\url{http://data.giss.nasa.gov/gistemp/tabledata/GLB.Ts.txt}
}
\usage{
gmst
}
\description{
A dataset containing the global mean surface temperature deviations for the
years 1881 through 2005 inclusive.
}
\details{
This version was downloaded in 2006 and has been superseded by updated
series based on more temperature measurements. See James Hodges, "Richly
Parameterized Linear Models" (2013) for details and analysis.
}
\keyword{datasets}

