% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{hmo}
\alias{hmo}
\title{HMO premiums}
\format{A data frame with 340 rows and 6 variables:
\describe{
  \item{state}{the state (two-letter abbreviation).}
  \item{plan}{a two-character code identifying the plan.}
  \item{indPrem}{total premium for an individual (dollars).}
  \item{famPrem}{total premium for a family (dollars).}
  \item{indEnr}{total enrollment of federal employees as individuals (number
  of persons).}
  \item{famEnr}{total enrollment of federal employees as famlilies (number of
  families).}
  }}
\usage{
hmo
}
\description{
A dataset containing has one row for each of 341 healthcare plans offered by
health maintenance organizations (HMOs). For state-level data, see
\code{\link{hmo_states}}.
}
\details{
This dataset was assembled in the mid-1990s to estimate the cost of moving families
from a Department of Defense healthcare plan to private plans. See James Hodges,
"Richly Parameterized Linear Models" (2013) for details and analysis.
}
\keyword{datasets}

