% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{hmo_states}
\alias{hmo_states}
\title{HMO states}
\format{A data frame with 45 rows and 4 variables:
\describe{
  \item{state}{the state (two-letter abbreviation).}
  \item{expPerAdm}{state average expenses per admission
  (dollars;  American Medical Association 1991 Annual Survey of Hospitals)}
  \item{pop}{population (1990 US Census)}
  \item{region}{region, from the Marion Merrill Dow Managed Care Digest
  1991. See details for abbreviations.}
  }}
\usage{
hmo_states
}
\description{
A dataset containing state- or jurisdiction-level information on the states
in which health maintainance organization (HMO) plans are offered. See
\code{\link{hmo}} for plan-level information.
}
\details{
Regional Abbreviations: MA = Middle Atlantic, MT = Mountain,
NC = North Central, NE = New England, PA = Pacific, SA = South Atlantic,
SC = South Central. See James Hodges, "Richly Parameterized Linear Models"
(2013) for details and analysis of this dataset.
}
\keyword{datasets}

