% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boxfuncs.R
\name{makebox}
\alias{makebox}
\title{Make a box with bounds}
\usage{
makebox(lims.sigsqe = NA, lims.sigsqs = NA, status = NA, lines)
}
\arguments{
\item{lims.sigsqe}{numeric vector of length 2 of the form \code{c(lower, upper)}
containing the x-limits of the box.}

\item{lims.sigsqs}{numeric vector of length 2 of the form \code{c(lower, upper)}.
containing the y-limits of the box.}

\item{status}{factor of length of \code{lines} denoting the location of the box
relative to each line. If the box has a parent, these correspond to the status
of that parent. The levels are \{above, below, straddle\}.}

\item{lines}{dataframe containing the constants that define the shape of each
  term in the sum. Output from \code{\link{findlines}}.}
}
\value{
A list containing:
\itemize{
  \item limits of the box in \eqn{\sigma^2_e} (numeric vector of length 2)
  \item limits of the box in \eqn{\sigma^2_s} (numeric vector of length 2)
  \item the status of the box relative to each line (factor of length of the
  number of terms in the sum)
  \item the bounds on the objective function (numeric vector of length 2 of
  the form \code{c(lower, upper)})
  }
}
\description{
Given the limits of the box, calculate its status relative to each line and
calculate the bounds on the objective function within that box.
}

