% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vizfuncs.R
\name{showboxes}
\alias{showboxes}
\title{Plot of box partition}
\usage{
showboxes(boxes)
}
\arguments{
\item{box}{a list of boxes, usually output from \code{\link{fitlmm}}.}
}
\description{
Generates a plot showing the partitioning of the parameter space
into boxes.
}

