% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vizfuncs.R
\name{showlines}
\alias{showlines}
\title{Plot of linear terms}
\usage{
showlines(lines)
}
\arguments{
\item{lines}{dataframe containing the constants that define the shape of each
term in the sum. Output from \code{\link{findlines}}.}
}
\description{
Generate a plot of the lines that correspond to where each term
in the sum of the objective function is maximized.
}

