% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boxfuncs.R
\name{splitbox}
\alias{splitbox}
\title{Split a parent box into 4 children.}
\usage{
splitbox(box, lines)
}
\arguments{
\item{box}{list containing the properties of a parent box. Created as output
from \code{\link{makebox}}.}

\item{lines}{dataframe containing the constants that define the shape of each
  term in the sum. Output from \code{\link{findlines}}.}
}
\value{
a list of the four child boxes.
}
\description{
Split a parent box into four equally sized boxes as part of the branching part
of the algorithm.
}

